+++ src/index.ts
@@ -25,6 +25,8 @@ export { default as IgcRadioComponent } from './components/radio/radio';
 export { default as IgcRadioGroupComponent } from './components/radio-group/radio-group';
 export { default as IgcRatingComponent } from './components/rating/rating';
 export { default as IgcRippleComponent } from './components/ripple/ripple';
+export { default as IgcSliderComponent } from './components/slider/slider';
+export { default as IgcRangeSliderComponent } from './components/slider/range-slider';
 export { default as IgcSwitchComponent } from './components/checkbox/switch';
 
 export { SizableMixin } from './components/common/mixins/sizable';